src: Move CUDA backend directory and use a separate CMake file for collecting its files
