+++ doc/Doxyfile.in
@@ -842,9 +842,8 @@ EXCLUDE_SYMLINKS       = NO
 # Note that the wildcards are matched against the file with absolute path, so to
 # exclude all test directories for example use the pattern */test/*
 
-EXCLUDE_PATTERNS       = */impl/*.cuh \
-                         */impl/*.h \
-                         */impl/*.hpp
+EXCLUDE_PATTERNS       = */impl/* \
+                         */cuda/*
 
 # The EXCLUDE_SYMBOLS tag can be used to specify one or more symbol names
 # (namespaces, classes, functions, etc.) that should be excluded from the
